+++ openedx_lti_tool_plugin/settings/common.py
@@ -44,7 +44,7 @@ def plugin_settings(settings: LazySettings):
     settings.OLTITP_ENABLE_LTI_TOOL = False
 
     # Username generation settings
-    settings.OLTITP_USERNAME_GENERATION_STRATEGY = 'uuid'  # 'uuid' | 'email_prefix'
+    settings.OLTITP_USERNAME_BASE_SOURCE = 'name'  # 'name' | 'email'
 
     # Deep linking settings
     settings.OLTITP_DEEP_LINKING_FORM_TEMPLATE = 'openedx_lti_tool_plugin/deep_linking/form.html'